src/test/javascript/spec/app/entities/segment/segment.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { SegmentService } from 'app/entities/segment/segment.service';
import { ISegment, Segment } from 'app/shared/model/segment.model';

describe('Service Tests', () => {
  describe('Segment Service', () => {
    let injector: TestBed;
    let service: SegmentService;
    let httpMock: HttpTestingController;
    let elemDefault: ISegment;
    let expectedResult: ISegment | ISegment[] | boolean | null;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(SegmentService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new Segment(0, 0, 0, 0, 0, 0, 0, 0);
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a Segment', () => {
        const returnedFromService = Object.assign(
          {
            id: 0,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.create(new Segment()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a Segment', () => {
        const returnedFromService = Object.assign(
          {
            start: 1,
            duration: 1,
            confidence: 1,
            loudnessStart: 1,
            loudnessMax: 1,
            loudnessMaxTime: 1,
            loudnessEnd: 1,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of Segment', () => {
        const returnedFromService = Object.assign(
          {
            start: 1,
            duration: 1,
            confidence: 1,
            loudnessStart: 1,
            loudnessMax: 1,
            loudnessMaxTime: 1,
            loudnessEnd: 1,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a Segment', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
